+++ pkg/handler/private_handler.go
@@ -1,0 +1,34 @@
+package handler
+
+import (
+	"context"
+
+	"github.com/instill-ai/pipeline-backend/pkg/datamodel"
+	"github.com/instill-ai/pipeline-backend/pkg/service"
+
+	pipelinePB "github.com/instill-ai/protogen-go/vdp/pipeline/v1alpha"
+)
+
+// PrivateHandler handles private API
+type PrivateHandler struct {
+	pipelinePB.UnimplementedPipelinePrivateServiceServer
+	service service.Service
+}
+
+// NewPrivateHandler initiates a handler instance
+func NewPrivateHandler(ctx context.Context, s service.Service) pipelinePB.PipelinePrivateServiceServer {
+	datamodel.InitJSONSchema(ctx)
+	return &PrivateHandler{
+		service: s,
+	}
+}
+
+// GetService returns the service
+func (h *PrivateHandler) GetService() service.Service {
+	return h.service
+}
+
+// SetService sets the service
+func (h *PrivateHandler) SetService(s service.Service) {
+	h.service = s
+}
